Refactored some more
