+++ backmap_pi/__init__.py
@@ -5,7 +5,7 @@ Backmapping for pi-conjugated peptides
 
 # Add imports here
 from .backmapping import *
-from .preprocessing import *
+from .topology import *
 
 # Handle versioneer
 from ._version import get_versions
